requirements.txt
attrs==21.4.0
numpy==1.23.1
Pillow==9.3.0
pyglet==1.5.26
python-utils==3.4.5
